Moves Link component to shared
