+++ src/components/Lesson/Lesson.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import StyledLink from '../Link';
+import StyledLink from '../shared/Link';
 import Block from '../shared/Block';
 import LessonButton, {PrimitiveLessonButton} from '../LessonButton';
 import ContentSection from '../shared/ContentSection';
@@ -26,7 +26,7 @@ const Lesson = ({lesson, slug}) => {
   return (
     <div className="Lesson">
       <div className="Lesson-homeButton">
-        <StyledLink variation={'tertiary'} path={`/courses/${course}/`}>
+        <StyledLink variation="tertiary" path={`/courses/${course}/`}>
           Back to {courseName}
         </StyledLink>
       </div>
@@ -82,7 +82,7 @@ const Lesson = ({lesson, slug}) => {
               <li className="Lesson-readingListItem" key={readingLink.title}>
                 <StyledLink
                   isExternal
-                  variation={'tertiary'}
+                  variation="tertiary"
                   path={readingLink.link}
                 >
                   {readingLink.title}
@@ -103,7 +103,7 @@ const Lesson = ({lesson, slug}) => {
           <StyledLink
             path={preReadQuiz !== null ? `${slug}/quiz` : preReadQuizLink}
             isExternal={preReadQuiz === null}
-            variation={'tertiary'}
+            variation="tertiary"
             lessonData={lesson}
           >
             Quiz Link
@@ -137,7 +137,7 @@ const Lesson = ({lesson, slug}) => {
         <StyledLink
           isExternal
           path="https://docs.google.com/forms/d/e/1FAIpQLSeiB_M1YmwwwG9BNhGnd1Nn_BhnzOfHFUDrZGz1PAvm8A1NxA/viewform"
-          variation={'tertiary'}
+          variation="tertiary"
         >
           Survey Link
         </StyledLink>
